refactor(coding): migrate tui playback testkit
