+++ docs/internals/architecture/harnesstui/README.md
@@ -12,9 +12,9 @@ direction is:
 `loushang.coding.presentation.tui.*` -> `loushang.harnesstui`
 `loushang.harnesstui` -> `loushang.harness`
 
-`loushang.coding.testing.tui` -> `loushang.harnesstui.testing`
-                              -> `loushang.harnesstui`
-                              -> `loushang.tui`
+`tests/coding/tui_support` -> `loushang.harnesstui.testing`
+                           -> `loushang.harnesstui`
+                           -> `loushang.tui`
 ```
 
 The reverse dependencies are forbidden. In particular, `loushang.harnesstui`
@@ -345,7 +345,7 @@ shared surface displays the prepared approval facts.
 
 `loushang.harnesstui.testing` is opt-in test support for exercising the
 product-neutral interaction ports above. Its dependency direction is
-`loushang.coding.testing.tui` -> `loushang.harnesstui.testing` ->
+`tests/coding/tui_support` -> `loushang.harnesstui.testing` ->
 `loushang.harnesstui` / `loushang.tui`. The reverse direction is forbidden:
 production Harnesstui must never import its testing package, and the generic
 TUI remains independent of both Harnesstui layers.
@@ -359,6 +359,8 @@ ports:
 - `loushang.harnesstui.testing.input_playback` owns decoded-input playback,
   neutral routed results, state snapshots, artifacts, and the fluent input
   scenario;
+- `loushang.harnesstui.testing.render_scenario` owns deterministic direct
+  rendering against a fake terminal and controllable clock;
 - `loushang.harnesstui.testing.screen_loop_playback` owns scripted TTY chunks,
   real screen-loop playback, captured output and state artifacts, and the
   fluent loop scenario;
@@ -373,14 +375,19 @@ ports:
   builders. They do not construct a product catalog at import time.
 
 These explicit modules are the stable testing entrypoints. The testing package
-initializer intentionally does not re-export them. Coding binds the neutral
-recipes into its concrete catalog under `loushang.coding.testing.tui.scenarios`
-and retains the app/router adapters, product-only scenarios, fakes, CLI runner,
-product copy, fixture volumes, and render-performance budgets. The temporary
+initializer intentionally does not re-export them. `ConversationRenderScenario`,
+the input and screen-loop drivers, and the scenario factory own the reusable
+fixture mechanics; `loushang.tui.playback_suite.run_playback_cli` owns catalog
+selection and reporting. Repository-local support in `tests/coding/tui_support`
+only binds those facilities into a concrete Coding catalog and retains
+product-only scenarios, fakes, copy, fixture volumes, and render-performance
+budgets. The repository manual runner is `scripts/run_tui_playback.py`;
+none of this product test support is part of the installed Coding package. The
+temporary
 `loushang.coding.ui.playback*` and `loushang.coding.ui.perf_probe` compatibility
 paths were retired after their consumers moved to the canonical testing
 packages. Persisted Coding Session materialization remains in
-`loushang.coding.testing.tui.performance`.
+`loushang.coding.presentation.tui.history`.
 
 ## Canonical Import Cutover
 
@@ -392,8 +399,11 @@ directly:
 - terminal settings and playback-suite primitives from `loushang.tui`;
 - neutral conversation, status, and performance support from
   `loushang.harnesstui`;
-- Coding playback catalogs, fakes, runners, scenarios, and persisted-session
-  loading from `loushang.coding.testing.tui`.
+- Coding playback catalogs, fakes, and scenarios from repository-local
+  `tests/coding/tui_support`, invoked through
+  `scripts/run_tui_playback.py`;
+- persisted-session transcript loading from
+  `loushang.coding.presentation.tui.history`.
 
 `loushang.coding.ui.cli` remains the product console entrypoint, while real
 Coding UI adapters that own product policy, copy, or runtime binding remain in
