+++ docs/internals/testing/screen-tui-playback.md
@@ -5,9 +5,10 @@ change can affect terminal behavior, not just pure component rendering.
 
 Reusable conversation drivers and product-neutral scenario recipes live under
 `loushang.harnesstui.testing`. The bound Coding suite, product fakes, frozen
-frame budgets, and CLI live under `loushang.coding.testing.tui`. The former
-`loushang.coding.ui.playback*` compatibility imports are retired; tests use the
-canonical testing packages directly.
+frame budgets, and product profiles are repository-local support under
+`tests/coding/tui_support`. They are not installed as a Coding compatibility
+package. The former `loushang.coding.ui.playback*` compatibility imports are
+retired; tests use the shared drivers and local support directly.
 
 Good candidates include:
 
@@ -67,9 +68,9 @@ smoke path.
 Useful direct smoke commands:
 
 ```bash
-uv --cache-dir .uv-cache run --extra dev python -m loushang.coding.testing.tui.runner composer-selection-stress --artifacts /tmp/loushang-selection-playback --include-frames
-uv --cache-dir .uv-cache run --extra dev python -m loushang.coding.testing.tui.runner product-composed-interaction --artifacts /tmp/loushang-product-playback --include-frames
-uv --cache-dir .uv-cache run --extra dev python -m loushang.coding.testing.tui.runner product-streaming-control-flow --artifacts /tmp/loushang-product-streaming-playback --include-frames
+uv --cache-dir .uv-cache run --extra dev python scripts/run_tui_playback.py composer-selection-stress --artifacts /tmp/loushang-selection-playback --include-frames
+uv --cache-dir .uv-cache run --extra dev python scripts/run_tui_playback.py product-composed-interaction --artifacts /tmp/loushang-product-playback --include-frames
+uv --cache-dir .uv-cache run --extra dev python scripts/run_tui_playback.py product-streaming-control-flow --artifacts /tmp/loushang-product-streaming-playback --include-frames
 ```
 
 For a public, product-neutral playback example, see
