+++ examples/tui/39_native_coding_long_transcript_perf.py
@@ -5,7 +5,9 @@ import asyncio
 import sys
 from pathlib import Path
 
-from loushang.coding.testing.tui.performance import load_session_history_records
+from loushang.coding.presentation.tui.history import (
+    load_persisted_session_history_records,
+)
 from loushang.coding.ui.screen_app import ScreenCodingTuiApp
 from loushang.harnesstui.testing.performance import (
     build_synthetic_long_transcript_records,
@@ -48,7 +50,9 @@ def main(argv: list[str] | None = None) -> int:
     args = parse_args(sys.argv[1:] if argv is None else argv)
 
     if args.session_file:
-        records = asyncio.run(load_session_history_records(Path(args.session_file)))
+        records = asyncio.run(
+            load_persisted_session_history_records(Path(args.session_file))
+        )
         source = str(Path(args.session_file).expanduser().resolve())
     else:
         records = build_synthetic_long_transcript_records(
