--- scripts/run_native_tui_playback.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-from loushang.coding.testing.tui.runner import main
-
-if __name__ == "__main__":
-    main()
