--- src/loushang/coding/testing/__init__.py
@@ -1,1 +1,0 @@
-"""Coding product test-support packages."""
