--- src/loushang/coding/testing/tui/__init__.py
@@ -1,1 +1,0 @@
-"""Coding TUI playback adapters and product scenarios."""
