+++ tests/coding/tui_support/__init__.py
@@ -1,0 +1,1 @@
+"""Repository-local support for Coding TUI integration tests."""
