+++ tests/coding/tui_support/scenarios/budgets.py
