+++ tests/coding/tui_support/scenarios/lifecycle.py
@@ -2,11 +2,11 @@
 
 from __future__ import annotations
 
-from loushang.coding.testing.tui.scenario_binding import (
+from loushang.harnesstui.testing.scenarios.lifecycle import lifecycle_scenarios
+from tests.coding.tui_support.scenario_binding import (
     CODING_SCENARIO_FACTORY,
     CODING_SCENARIO_FRAME_CONTRACTS,
 )
-from loushang.harnesstui.testing.scenarios.lifecycle import lifecycle_scenarios
 
 LIFECYCLE_SCENARIOS = lifecycle_scenarios(
     CODING_SCENARIO_FACTORY,
