+++ tests/coding/tui_support/scenarios/terminal.py
@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-from loushang.coding.testing.tui.scenario_binding import CODING_SCENARIO_FACTORY
 from loushang.harnesstui.testing.scenarios.terminal import terminal_scenarios
+from tests.coding.tui_support.scenario_binding import CODING_SCENARIO_FACTORY
 
 TERMINAL_SCENARIOS = terminal_scenarios(CODING_SCENARIO_FACTORY)
